# go_app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Agendly",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/get-shop-by-user": {
            "get": {
                "description": "Retrieve all shop by user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "GetAllShopByUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id_user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/get-shop/{id_kind}/{city}": {
            "get": {
                "description": "Retrieve all shop by kind and city",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "GetAllShopByKindAndCity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "shop city",
                        "name": "city",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "kind id_kind",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/new-shop": {
            "post": {
                "description": "create a shop and its informations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "AddShop",
                "parameters": [
                    {
                        "description": "shop name",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop zip_code",
                        "name": "zip_code",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop city",
                        "name": "city",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop lat",
                        "name": "lat",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop long",
                        "name": "long",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop country",
                        "name": "country",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop phone",
                        "name": "phone",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop description",
                        "name": "description",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "kind id_kind",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "user id_user",
                        "name": "id_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/shop/{id_shop}": {
            "delete": {
                "description": "Delete a shop by its id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "DeleteShop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "shop id_shop",
                        "name": "id_shop",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/shop/{id}": {
            "get": {
                "description": "Retrieve all shop by id shope",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "GetShop",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "shop id_shop",
                        "name": "shop",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "description": "Update a shop by its id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "UpdateShop",
                "parameters": [
                    {
                        "description": "shop name",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop zip_code",
                        "name": "zip_code",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop city",
                        "name": "city",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop lat",
                        "name": "lat",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop long",
                        "name": "long",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop country",
                        "name": "country",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop phone",
                        "name": "phone",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "shop description",
                        "name": "description",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/shops/nearby/{lng}/{lat}/{kind}": {
            "get": {
                "description": "Retrieve all shop near the area defined by latitude and longitude",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "GetAllShopNear",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "shop long",
                        "name": "lng",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "shop lat",
                        "name": "lat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "kind id_kind",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}